Splitting out asserts from app for maintain and read ability
