+++ src/test/java/integration/AssertStateIT.java
@@ -27,7 +27,7 @@ public class AssertStateIT extends Selenified {
         // use this object to manipulate the app
         App app = this.apps.get();
         // perform the verification
-        app.titleEquals("Selenified Test Page");
+        app.azzert().titleEquals("Selenified Test Page");
         // perform the verification
         finish();
     }
